rm depends on audio-api
The Soundboard is designed to be run in two modes,
one for editing the Songs, and the other for using them. The primary mode is editing, and so
it does not need to audio-api to be active on startup. If that is needed, 'docker compose up'
will build the full service.

This change allows a targetted 'docker compose up soundboard-ui -d'.